Move source files into a subfolder
