+++ Source/RotaryWingAircraft/Private/RWA/Util.h
