+++ Source/RotaryWingAircraft/Public/RWA/AnimNode_RotorController.h
