+++ Source/RotaryWingAircraft/Public/RWA/InputModifier_Dampen.h
